feat: add InGame (not completed)

+++ Assets/Scripts/Constants.cs
@@ -6,8 +6,8 @@ public class Constants
     public const int MAP_SIZE = 10;
     public const int CELL_SIZE = 10;
     public const int CELL_ELEVATION = 5;
-    public const int LEFT_OFFSET = -100;
-    public const int BOTTOM_OFFSET = -55;
+    public const float LEFT_OFFSET = -100.0f;
+    public const float BOTTOM_OFFSET = -55.0f;
     public const string CELL_COLOR_NORMAL = "#B9F6FFFF";
     public const string CELL_COLOR_VALID = "#00E62EFF";
     public const string CELL_COLOR_INVALID = "#E64100FF";
